Cleanup - moved CreateCommentRequest to Comments DTOs (for some reason it was in venues)
