+++ HobbyMatch.BL/DTOs/Comments/CreateCommentRequest.cs
@@ -1,5 +1,5 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace HobbyMatch.Domain.Requests;
+namespace HobbyMatch.BL.DTOs.Comments;
 
 public record CreateCommentRequest(int EventId, [MaxLength(300)] string Content);
